* commit 'bdf1f41b7d26e4d07f3e121ed2f973dd8e8f1433':
  disabled oracle in local. enabled oracle for testing in dev
  changed info to debug
  updated readme file
  updated readme file
  multi-db support added functionality to change setting through command line
  multi-db support added functionality to change setting through command line
  multi-db support (disabled cassandra for now)
  multi-db support
  multi-db support
  multi-db support
  multi-db support
  multi-db support
  multi-db support
  multi-db support
  removed import
  cassandra implementation (final).
  cassandra implementation
  cassandra implementation (initial commit).
  cassandra implementation

+++ cakeshop-api/src/main/java/com/jpmorgan/ib/caonpd/cakeshop/dao/WalletDAO.java
@@ -1,8 +1,10 @@
 package com.jpmorgan.ib.caonpd.cakeshop.dao;
 
 import com.jpmorgan.ib.caonpd.cakeshop.model.Account;
+import java.util.ArrayList;
 
 import java.util.List;
+import org.hibernate.Criteria;
 
 import org.hibernate.Session;
 import org.springframework.stereotype.Repository;
@@ -13,19 +15,27 @@ import org.springframework.transaction.annotation.Transactional;
 public class WalletDAO extends BaseDAO {
 
     public List<Account> list() {
-        return hibernateTemplate.loadAll(Account.class);
+        if (null != getCurrentSession()) {
+            Criteria criteria = getCurrentSession().createCriteria(Account.class);
+            return criteria.list();
+        }
+        return new ArrayList();
     }
 
     public void save(Account account) {
-        this.hibernateTemplate.save(account);
+        if (null != getCurrentSession()) {
+            getCurrentSession().save(account);
+        }
     }
 
     @Override
     public void reset() {
-        Session session = getCurrentSession();
-        session.createSQLQuery("DELETE FROM ACCOUNTS").executeUpdate();
-        session.flush();
-        session.clear();
+        if (null != getCurrentSession()) {
+            Session session = getCurrentSession();
+            session.createSQLQuery("DELETE FROM ACCOUNTS").executeUpdate();
+            session.flush();
+            session.clear();
+        }
     }
 
 }
